refactor: implement authentication module and update routing to use new structure
